Move rendering from miniscreen to assistant
